+++ src/project/repository/FlightRepository.java
@@ -2,5 +2,5 @@ package project.repository;
 
 import project.entity.Flight;
 
-public interface FlightRepository extends CrudRepository<Flight, Integer> {
+public interface FlightRepository extends CrudRepository<Flight, Long> {
 }
